third iteration
Changes in the structure involve making Stage a HashSet instead of an object, changing add and remove to be functions of Gitlet instead of the removed Stage class. Commit has also been heavily edited from an Object with basic features to appropriate data structures (particularly maps and trees) that are compatible with staging operations. Initialise was also changed from a linked list to a tree for more flexibility. More directions have been written out for init as well.